pyproject.toml
[project]
name = "qcnn-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "hydra-core>=1.3.2",
    "ipython>=9.9.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.1",
    "python-dotenv>=1.2.1",
    "qiskit>=2.3.0",
    "qiskit-aer>=0.17.2",
    "qiskit-algorithms>=0.4.0",
    "qiskit-machine-learning>=0.9.0",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
]

[dependency-groups]
dev = [
    "pre-commit>=4.0.1",
    "ruff>=0.9.2",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I"] # Basic errors, formatting, and import sorting
fixable = ["ALL"]
